refactor(game-service): change file naming and names in the code
