+++ services/game_service/src/services/rabbitMQ-client.ts
